[별희] ItemImage에서 ItemId uint32로 변경

+++ Source/MetaAuction/UI/MAWidgetHelperLibrary.h
@@ -17,5 +17,5 @@ class METAAUCTION_API UMAWidgetHelperLibrary : public UBlueprintFunctionLibrary
 public:
 
 	UFUNCTION(BlueprintCallable, Category = Widget)
-	static void RequestImageByItemID(class UImage* ImageWidget, uint8 InItemID, uint8 InImageNumber = 1);
+	static void RequestImageByItemID(class UImage* ImageWidget, uint32 InItemID, uint8 InImageNumber = 1);
 };
